Bump to python 3.14 and fix a few small bugs
