+++ archive/lambdas/ftpToUkmon/pythoncode/Dockerfile
@@ -1,10 +1,9 @@
-FROM public.ecr.aws/lambda/python:3.8
+FROM public.ecr.aws/lambda/python:3.14
 
 COPY requirements.txt ./
-RUN python3.8 -m pip install -r requirements.txt -t .
-RUN yum update -y
+RUN python -m pip install -r requirements.txt -t .
 ENV LD_LIBRARY_PATH ./
 COPY *.py ./
 COPY *.csv ./
 
-CMD ["ftpToUkmon1.lambda_handler"]
+CMD ["ftpToUkmon.lambda_handler"]
