+++ archive/lambdas/ftpToUkmon/pythoncode/requirements.txt
@@ -1,3 +1,3 @@
-numpy
 scipy
 pytz
+numpy
