+++ archive/lambdas/ftpToUkmon/template.yaml
@@ -34,4 +34,5 @@ Resources:
     Metadata:
       Dockerfile: Dockerfile
       DockerContext: ./pythoncode
-      DockerTag: python3.8-v1
+      DockerTag: python3.14-v1
+      DockerImage: ftpToUkmon
