+++ archive/lambdas/ftpToUkmon/tests/localTest.ps1
@@ -1,7 +1,7 @@
 # test the lambda
 
-sam build --profile ukmonshared
-sam local invoke --profile ukmonshared -e ./tests/testEvent.json
+Remove-Item ./tests/new_data.txt
+sam local invoke --profile ukmda_admin -e ./tests/testEvent.json
 start-sleep 1
-aws s3 cp s3://ukmda-shared/matches/single/new/ukmon_UK0006_20221122_164325_836988.csv ./tests/new_data.txt
+aws s3 cp s3://ukmda-shared/matches/single/new/ukmda_UK0006_20240811_201327_827902.csv ./tests/new_data.txt --profile ukmda_admin
 bash -c "diff ./tests/expected_results.txt ./tests/new_data.txt"
