+++ archive/lambdas/vidUploader/template.yaml
@@ -16,7 +16,7 @@ Resources:
       FunctionName: videohandler
       CodeUri: pythoncode/
       Handler: app.lambda_handler
-      Runtime: python3.11
+      Runtime: python3.14
       Timeout: 900
       MemorySize: 512
       EphemeralStorage:
